requirements.txt
mediapipe==0.10.21
opencv-python==4.8.1.78
numpy==1.24.3
azure-cognitiveservices-speech==1.40.0
python-dotenv==1.0.0
scipy==1.11.4